Support multi-byte VarInts throughout (#128)
* refactor: start prep for no_std version

* refactor: remove BE functions from BTCUtils and Parse functions from ValidateSPV in all languages

* feature: support multi-byte VarInt in vin and vout

* tests: update vectors for multi-byte VarInt

* tests: add tests for parseVarInt

* feature: multi-byte VarInt support in Rust

* tests: parse_var_int tests in Rust

* tests: mark uncovered lines in solidity and rust

* bug: delete removed method declarations from header file

* test: add missing error match

* feature: multi-byte VarInt support in Go

* chore: update example CLI code with new function signatures

* feature: multi-byte VarInts in JS

* feature: multi-byte VarInt parsing for C

* feature: multi-byte VarInt in extract input/output for rust

* feature: multi-byte VarInt in extract input/output for solidity

* feature: multi-byte VarInt in extract input/output for JS

* feature: multi-byte VarInt in extract input/output for golang

* feature: multi-byte VarInt in extract input/output for C

* chore: add coverage TODOs to C files

* added more test coverage

* 100% coverage in golang, checking other languages

* js - 100% coverage

* python: delete be function

* tests: rust coverage on btcspv back to 100%

* tests: lots of coverage and adjustments. this one got away from me.

* tests: add notes to c swap demo about why it's failing currently

* tests: add note to swap-demo about where to load args

* tests: full coverage on everything we care about

* chore: remove outdated coverage notes

* feature: namespace utility functions

* chore: hackathon ideas in C Readme

* chore: more README instructions

+++ golang/cli/vout.go
@@ -79,7 +79,7 @@ func ParseVout(vout []byte) string {
 	var formattedOutputs string
 	for i := 0; i < numOutputs; i++ {
 		// Extract each vout at the specified index
-		vout, err := btcspv.ExtractOutputAtIndex(vout, uint8(i))
+		vout, err := btcspv.ExtractOutputAtIndex(vout, uint(i))
 		if err != nil {
 			return fmt.Sprintf("%s\n", err)
 		}
@@ -129,7 +129,6 @@ func parseOutput(output []byte) (uint, OutputType, []byte) {
 
 	return value, outputType, payload
 }
-
 
 // GetOutputType returns the name of the output type associated with the number
 func GetOutputType(outputType OutputType) string {
